+++ src/arch/risc-v/kernel/sbi.c
@@ -1,8 +1,8 @@
 #include "defs.h"
 
-void consolePutchar(u64 c) { SBI_ECALL_1(SBI_CONSOLE_PUTCHAR, c); }
+void consolePutchar(usize c) { SBI_ECALL_1(SBI_CONSOLE_PUTCHAR, c); }
 
-u64 consoleGetchar() { return SBI_ECALL_0(SBI_CONSOLE_GETCHAR); }
+usize consoleGetchar() { return SBI_ECALL_0(SBI_CONSOLE_GETCHAR); }
 
 void shutdown() {
   SBI_ECALL_0(SBI_SHUTDOWN);
